+++ packages/sdk/src/EventSenderEngine.test.ts
@@ -41,7 +41,7 @@ describe('EventSenderEngine unit tests', () => {
     expect(flushSpy).toHaveBeenCalledTimes(1);
     await jest.runAllTimersAsync();
   });
-  it('payload should prioritize message fields', async () => {
+  it('should store context under a context field ', async () => {
     const noBatchEngine = new EventSenderEngine({
       clientSecret: 'my_secret',
       maxBatchSize: 1,
@@ -52,7 +52,7 @@ describe('EventSenderEngine unit tests', () => {
       logger: {},
     });
     const uploadSpy = jest.spyOn(noBatchEngine, 'upload');
-    noBatchEngine.send({ a: 2, message: 3 }, 'my_event', { a: 0, message: 1 });
+    noBatchEngine.send({ value: 2 }, 'my_event', { message: 1 });
     await jest.runAllTimersAsync();
     expect(uploadSpy).toHaveBeenCalledTimes(1);
     expect(uploadSpy).toHaveBeenCalledWith({
@@ -63,12 +63,30 @@ describe('EventSenderEngine unit tests', () => {
           eventDefinition: 'my_event',
           eventTime: expect.any(String),
           payload: {
-            a: 0,
+            context: { value: 2 },
             message: 1,
           },
         },
       ],
     });
+  });
+  it('should throw if data contains a context field', async () => {
+    const noBatchEngine = new EventSenderEngine({
+      clientSecret: 'my_secret',
+      maxBatchSize: 1,
+      flushTimeoutMilliseconds: FLUSH_TIMEOUT,
+      fetchImplementation: mockFetch as any,
+      region: 'eu',
+      maxOpenRequests: MAX_OPEN_REQUESTS,
+      logger: {},
+    });
+    const uploadSpy = jest.spyOn(noBatchEngine, 'upload');
+    expect(() => {
+      // @ts-expect-error
+      noBatchEngine.send({ value: 2 }, 'my_event', { context: 0, message: 1 });
+    }).toThrow('Event data must not contain a context field');
+    await jest.runAllTimersAsync();
+    expect(uploadSpy).toHaveBeenCalledTimes(0);
   });
   it('should handle a lot of events', async () => {
     const eventCount = BATCH_SIZE * MAX_OPEN_REQUESTS * 2;