+++ packages/sdk/src/flags.ts
@@ -1,5 +1,4 @@
 import { Contextual } from '.';
-// import { AccessiblePromise } from './AccessiblePromise';
 import { Value } from './Value';
 
 export namespace FlagEvaluation {
@@ -41,5 +40,7 @@ export interface FlagResolver extends Contextual<FlagResolver> {
 
   subscribe(onStateChange?: StateObserver): () => void;
 
-  getFlag<T extends Value>(path: string, defaultValue: T): FlagEvaluation<T>;
+  evaluateFlag<T extends Value>(path: string, defaultValue: T): FlagEvaluation<Value.Widen<T>>;
+
+  getFlag<T extends Value>(path: string, defaultValue: T): Promise<Value.Widen<T>>;
 }