+++ packages/openfeature-server-provider/CHANGELOG.md
@@ -12,6 +12,25 @@
   * dependencies
     * @spotify-confidence/sdk bumped from 0.0.2 to 0.0.3
 
+## [0.2.5](https://github.com/spotify/confidence-sdk-js/compare/openfeature-server-provider-v0.2.4...openfeature-server-provider-v0.2.5) (2024-05-28)
+
+
+### 🐛 Bug Fixes
+
+* npm deployment ([#146](https://github.com/spotify/confidence-sdk-js/issues/146)) ([0c93173](https://github.com/spotify/confidence-sdk-js/commit/0c931732a8c8df4b73d5e7a5b3bcda21684cb441))
+
+## [0.2.4](https://github.com/spotify/confidence-sdk-js/compare/openfeature-server-provider-v0.2.3...openfeature-server-provider-v0.2.4) (2024-05-27)
+
+
+### 🐛 Bug Fixes
+
+* widen flag return types ([#142](https://github.com/spotify/confidence-sdk-js/issues/142)) ([6554e8c](https://github.com/spotify/confidence-sdk-js/commit/6554e8c83c6c49103f11fbdcf3f53c5576870788))
+
+
+### ✨ New Features
+
+* confidence flag api ([#141](https://github.com/spotify/confidence-sdk-js/issues/141)) ([3583415](https://github.com/spotify/confidence-sdk-js/commit/3583415957915a4d181316b66e5549071836799f))
+
 ## [0.2.3](https://github.com/spotify/confidence-openfeature-provider-js/compare/openfeature-server-provider-v0.2.2-rc.0...openfeature-server-provider-v0.2.3) (2024-05-03)